Timmer, Pomodoro Integrated

+++ backend/app/focus/models.py
@@ -4,16 +4,16 @@ Modes: pomodoro | flowtime | deepwork | break
 Status: pending → running → completed | cancelled | interrupted
 is_break=True sessions are excluded from focus time analytics.
 """
-from sqlalchemy import BigInteger, String, Integer, DateTime, Boolean, Float, ForeignKey
+from sqlalchemy import Integer, String, Integer, DateTime, Boolean, Float, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy.sql import func
 from app.database import Base
 
 class FocusSession(Base):
     __tablename__ = "focus_sessions"
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True)
-    task_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("tasks.id", ondelete="SET NULL"), nullable=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    user_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True)
+    task_id: Mapped[int] = mapped_column(Integer, ForeignKey("tasks.id", ondelete="SET NULL"), nullable=True)
     mode: Mapped[str] = mapped_column(String(20), nullable=False)
     status: Mapped[str] = mapped_column(String(20), default="pending")
     started_at: Mapped[DateTime] = mapped_column(DateTime(timezone=True), nullable=True)
