+++ backend/app/models/channel.py
@@ -2,15 +2,15 @@
 models/channel.py — Channel Entity
 For tagging tasks (e.g., #work, #personal)
 """
-from sqlalchemy import BigInteger, String, ForeignKey, DateTime
+from sqlalchemy import Integer, String, ForeignKey, DateTime
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy.sql import func
 from app.database import Base
 
 class Channel(Base):
     __tablename__ = "channels"
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    user_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True)
     name: Mapped[str] = mapped_column(String(50), nullable=False)
     color: Mapped[str] = mapped_column(String(20), default="#10B981") # Default green
     created_at: Mapped[DateTime] = mapped_column(DateTime(timezone=True), server_default=func.now())
